move context to its own folder
